flask_model_api/requirements.txt
numpy
tensorflow
Pillow
flask
gunicorn